requirements.txt
numpy==1.19.2
requests==2.24.0
Flask==1.1.2
Pillow==8.1.0
gunicorn